Create Greed
Greed is a game in which you roll a dice and accumulate score through the numbers that you roll. You can choose to keep rolling and accumulating more points in a turn, but if at any point, you roll either die on 'one' your turn immediately ends and you add zero points to your total score. In this program you verse an 'A.I.' (not really) and the first to 100 points wins.